refactor: comments, rename folders and files
